+++ src/ducks/Studio/AdvancedView/SocialContext.js
@@ -2,12 +2,10 @@ import React, { useState, useEffect } from 'react'
 import cx from 'classnames'
 import Icon from '@santiment-network/ui/Icon'
 import Calendar from './Calendar'
-import UsageTip from './UsageTip'
 import WordCloud from '../../../components/WordCloud/WordCloud'
 import TrendsTable from '../../../components/Trends/TrendsTable/TrendsTable'
 import GetHypedTrends from '../../../components/Trends/GetHypedTrends'
 import { INTERVAL_ALIAS } from '../../SANCharts/IntervalSelector'
-import { useDebounceEffect } from '../../../hooks'
 import { parseIntervalString, ONE_MONTH_IN_MS } from '../../../utils/dates'
 import styles from './SocialContext.module.scss'
